update funfoldes plugin
